requirements.txt
google-cloud-datastore==1.6.0
google-cloud-storage==1.8.0
google-cloud-vision==0.30.1
ImageHash==4.0
numpy==1.14.0
pandas==0.22.0
Pillow==5.0.0